requirements.txt
aiohttp>=3.9.0
feedparser>=6.0.10
python-dotenv>=1.0.0
pytz>=2024.1
